Adding new features

+++ src/app/repository/AccountRepository.ts
@@ -1,30 +1,81 @@
 import IRepository from "../interface/IRepository";
-import { Account } from "../types/Account";
 
 export default class AccountRepository implements IRepository<Account> {
 
-    load(id: any): Promise<Account>
+    getById(id: string): Promise<Account> 
     {
-        return new Promise( (resolve) => {
-            let retValue:Account = null;
+        return new Promise( (resolve, reject) => {
+            try
+            {
+                let retValue:Account = null;
 
-            if(typeof id == "string" && window.localStorage.getItem(id) != null) {
-                retValue = JSON.parse(window.localStorage.getItem(id));
+                if(typeof id == "string" && window.localStorage.getItem(id) != null) 
+                {
+                    retValue = JSON.parse(window.localStorage.getItem(id));
+                }
+
+                resolve(retValue);
             }
-
-            resolve(retValue);
+            catch(e)
+            {
+                reject(e);
+            }
         })
     }
 
-    save(value: Account): Promise<Account> 
+    create(value: Account): Promise<Account>
     {
-        return new Promise( (resolve) => {
-            window.localStorage.setItem(value.email, JSON.stringify(value));
-            resolve(value);
+        return new Promise( (resolve, reject) => {
+            try
+            {
+                if(!this.exists(value.email))
+                {
+                    window.localStorage.setItem(value.email, JSON.stringify(value));
+                    resolve(value);
+                }
+                else
+                {
+                    reject("Account already exists.")
+                }
+            }
+            catch(e) 
+            {
+                reject(e);
+            }
         });
     }
 
-    exists(id: any): boolean {
+    delete(value: Account): Promise<Account> {
+        return new Promise( (resolve, reject) => 
+        {
+            if(this.exists(value.email))
+            {
+                window.localStorage.removeItem(value.email);
+                resolve(value);
+            }
+            else
+            {
+                reject("Account does not exists.")
+            }
+        });
+    }
+
+    update(value: Account): Promise<Account> {
+        return new Promise( (resolve, reject) => 
+        {
+            if(this.exists(value.email)) 
+            {
+                window.localStorage.setItem(value.email, JSON.stringify(value));
+                resolve(value);
+            }
+            else
+            {
+                reject("Account does not exists.")
+            }
+        });
+    }
+
+    exists(id: string): boolean {
         return window.localStorage.getItem(id) != null;
     }
 }
